Get test working with test db url
